Proper way to show Spinner in Toolbar from the Fragment
